createEditQuery

+++ src/pages/links.ts
@@ -3,7 +3,8 @@ import { LinksQuery } from '~/domain/queries'
 import { DeleteLinkCommand, UpdateLinksCommand } from '~/domain/commands'
 import { ILink } from '~/domain/models'
 import { Getter } from 'vuex-class'
-import { CreateEditCommand } from '~/domain/commands/createEdit.command'
+import { CreateEditQuery } from '~/domain/queries/createEdit.query'
+import { ConfirmWindowQuery } from '~/domain/queries/confirmWindow.query'
 
 export default class Links extends Vue {
   @Getter('links/getLinks') links: Array<ILink>
@@ -24,19 +25,21 @@ export default class Links extends Vue {
   async edit(id: string) {
     const found = this.links.find(link => link.id === id)
     if (found) {
-      const command = new CreateEditCommand({
+      const query = new CreateEditQuery<ILink>({
         component: 'create-edit-link',
         componentProps: {
-          id,
-          url: found.url,
-          name: found.name
+          item: {
+            id,
+            url: found.url,
+            name: found.name
+          }
         },
         modal: {
           title: 'Edit link',
           width: '30%'
         }
       })
-      const result = await this.$app.$commandBus.do<CreateEditCommand<ILink>, ILink>(command)
+      const result = await this.$app.$queryBus.exec<CreateEditQuery<ILink>, ILink>(query)
       if (!result) {
         return
       }
@@ -46,6 +49,12 @@ export default class Links extends Vue {
   }
 
   async remove(id: string) {
+    const isConfirm = await this.$app.$queryBus.exec(new ConfirmWindowQuery(
+      'Do you want to remove link?'
+    ))
+    if (!isConfirm) {
+      return
+    }
     try {
       await this.$app.$commandBus.do<DeleteLinkCommand, void>(new DeleteLinkCommand(id))
       await this.$app.$queryBus.exec<LinksQuery, Array<ILink>>(new LinksQuery())
